Fixed a bug in the grimm's quest for the gatherers key game dll. more cleanups. possible bug fixed in radiant.

+++ Doom3/Extras/ModSources/grimm/game/gamesys/Event.cpp
@@ -605,7 +605,7 @@ void idEvent::Restore( idRestoreGame *savefile ) {
 		savefile->ReadObject( event->object );
 		// read the args
 		savefile->ReadInt( argsize );
-		if( argsize != event->eventdef->GetArgSize() ) {
+		if( argsize != static_cast<int>( event->eventdef->GetArgSize() ) ) {
 			savefile->Error( "idEvent::Restore: arg size (%d) doesn't match saved arg size(%d) on event '%s'", event->eventdef->GetArgSize(), argsize, event->eventdef->GetName() );
 		}
 		if( argsize ) {
